# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MoneyMind AI - Personal Finance Intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  # ── Transactions ──────────────────────────────────────────────────────────
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [expense, income]
        - name: category_id
          in: query
          schema:
            type: integer
        - name: start_date
          in: query
          schema:
            type: string
        - name: end_date
          in: query
          schema:
            type: string
        - name: payment_method
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
    put:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Categories ────────────────────────────────────────────────────────────
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List all financial goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a financial goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    put:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Reminders ─────────────────────────────────────────────────────────────
  /reminders:
    get:
      operationId: listReminders
      tags: [reminders]
      summary: List all reminders
      parameters:
        - name: upcoming_days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create a reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderInput"
      responses:
        "201":
          description: Created reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{id}:
    put:
      operationId: updateReminder
      tags: [reminders]
      summary: Update a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderUpdate"
      responses:
        "200":
          description: Updated reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
    delete:
      operationId: deleteReminder
      tags: [reminders]
      summary: Delete a reminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Profile:
      type: object
      required: [id, full_name, currency, theme]
      properties:
        id:
          type: integer
        full_name:
          type: string
        occupation:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        job_status:
          type: ["string", "null"]
        income_type:
          type: ["string", "null"]
        currency:
          type: string
        country:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        monthly_income:
          type: ["number", "null"]
        salary_frequency:
          type: ["string", "null"]
        monthly_goal:
          type: ["number", "null"]
        weekly_savings_goal:
          type: ["number", "null"]
        emergency_fund_goal:
          type: ["number", "null"]
        week_start_day:
          type: ["string", "null"]
        theme:
          type: string
        date_format:
          type: ["string", "null"]
        avatar_url:
          type: ["string", "null"]
        notification_preferences:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
    ProfileInput:
      type: object
      required: [full_name, currency, theme]
      properties:
        full_name:
          type: string
        occupation:
          type: string
        company:
          type: string
        job_status:
          type: string
        income_type:
          type: string
        currency:
          type: string
        country:
          type: string
        state:
          type: string
        city:
          type: string
        monthly_income:
          type: number
        salary_frequency:
          type: string
        monthly_goal:
          type: number
        weekly_savings_goal:
          type: number
        emergency_fund_goal:
          type: number
        week_start_day:
          type: string
        theme:
          type: string
        date_format:
          type: string
        avatar_url:
          type: string
        notification_preferences:
          type: string

    # ── Transaction ───────────────────────────────────────────────────────────
    Transaction:
      type: object
      required: [id, type, amount, date, category_id]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [expense, income]
        amount:
          type: number
        date:
          type: string
        time:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        category_id:
          type: integer
        category_name:
          type: ["string", "null"]
        category_color:
          type: ["string", "null"]
        category_icon:
          type: ["string", "null"]
        subcategory:
          type: ["string", "null"]
        payment_method:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        is_recurring:
          type: boolean
        recurring_frequency:
          type: ["string", "null"]
        mood:
          type: ["string", "null"]
        need_or_want:
          type: ["string", "null"]
        is_business:
          type: boolean
        is_tax_deductible:
          type: boolean
        receipt_url:
          type: ["string", "null"]
        income_source:
          type: ["string", "null"]
        created_at:
          type: string
    TransactionInput:
      type: object
      required: [type, amount, date, category_id]
      properties:
        type:
          type: string
          enum: [expense, income]
        amount:
          type: number
        date:
          type: string
        time:
          type: string
        description:
          type: string
        category_id:
          type: integer
        subcategory:
          type: string
        payment_method:
          type: string
        location:
          type: string
        tags:
          type: string
        notes:
          type: string
        is_recurring:
          type: boolean
        recurring_frequency:
          type: string
        mood:
          type: string
        need_or_want:
          type: string
        is_business:
          type: boolean
        is_tax_deductible:
          type: boolean
        receipt_url:
          type: string
        income_source:
          type: string
    TransactionUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [expense, income]
        amount:
          type: number
        date:
          type: string
        time:
          type: string
        description:
          type: string
        category_id:
          type: integer
        subcategory:
          type: string
        payment_method:
          type: string
        location:
          type: string
        tags:
          type: string
        notes:
          type: string
        is_recurring:
          type: boolean
        recurring_frequency:
          type: string
        mood:
          type: string
        need_or_want:
          type: string
        is_business:
          type: boolean
        is_tax_deductible:
          type: boolean
        receipt_url:
          type: string
        income_source:
          type: string
    TransactionListResponse:
      type: object
      required: [transactions, total]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer

    # ── Category ──────────────────────────────────────────────────────────────
...
    Goal:
      type: object
      required: [id, name, target_amount, current_amount]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        target_amount:
          type: number
        current_amount:
          type: number
        target_date:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        is_completed:
          type: boolean
        created_at:
          type: string
    GoalInput:
      type: object
      required: [name, target_amount]
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        target_amount:
          type: number
        current_amount:
          type: number
        target_date:
          type: string
        description:
          type: string
    GoalUpdate:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        target_amount:
          type: number
        current_amount:
          type: number
        target_date:
          type: string
        description:
          type: string
        is_completed:
          type: boolean

    # ── Reminder ──────────────────────────────────────────────────────────────
    Reminder:
      type: object
      required: [id, title, due_date]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        due_date:
          type: string
        amount:
          type: ["number", "null"]
        category:
          type: ["string", "null"]
        is_recurring:
          type: boolean
        recurring_frequency:
          type: ["string", "null"]
        is_paid:
          type: boolean
        notify_days_before:
          type: integer
        created_at:
          type: string
    ReminderInput:
      type: object
      required: [title, due_date]
      properties:
        title:
          type: string
        description:
          type: string
        due_date:
          type: string
        amount:
          type: number
        category:
          type: string
        is_recurring:
          type: boolean
        recurring_frequency:
          type: string
        notify_days_before:
          type: integer
    ReminderUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        due_date:
          type: string
        amount:
          type: number
        category:
          type: string
        is_recurring:
          type: boolean
        recurring_frequency:
          type: string
        is_paid:
          type: boolean
        notify_days_before:
          type: integer

    # ── Dashboard ─────────────────────────────────────────────────────────────
...